+++ CHANGELOG.md
@@ -65,6 +65,7 @@ The following lists show the changes to the library grouped by domain.
 * fixing [`ally.maintain.disabled`][ally/maintain/disabled] to properly handle `tabindex` attribute changes
 * fixing [`ally.maintain.disabled`][ally/maintain/disabled] to properly disengage within ShadowHosts - [issue #107](https://github.com/medialize/ally.js/issues/107), [PR #108](https://github.com/medialize/ally.js/pull/108)
 * fixing [`ally.get.parents`][ally/get/parents] to resolve ancestry for `SVGElement` in Internet Explorer
+* adding [`ally.query.shadowHosts`][ally/query/shadow-hosts] to find elements hosting `ShadowRoot`s - [issue #110](https://github.com/medialize/ally.js/issues/110)
 
 #### Internals
 
@@ -278,6 +279,7 @@ Version `1.0.0` is a complete rewrite from the the early `0.0.x` releases, there
 [ally/observe/interaction-type]: http://allyjs.io/api/observe/interaction-type.html
 [ally/query/first-tabbable]: http://allyjs.io/api/query/first-tabbable.html
 [ally/query/focusable]: http://allyjs.io/api/query/focusable.html
+[ally/query/shadow-hosts]: http://allyjs.io/api/query/shadow-hosts.html
 [ally/query/tabbable]: http://allyjs.io/api/query/tabbable.html
 [ally/query/tabsequence]: http://allyjs.io/api/query/tabsequence.html
 [ally/style/focus-source]: http://allyjs.io/api/style/focus-source.html